+++ changelog.md
@@ -98,4 +98,4 @@
 - Added tutorial on how to change about page (user generated content)
 - Generate statistics page less frequently per default (stale age) [GH-14]
 - Performance update: reduced generation time by 50% [GH-14]
-
+- Added thunderstorm logic and icon to forecast table